requirements.txt
<<<<<<< HEAD
asgiref==3.5.2
Authlib==1.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.1
Django==4.0.5
django-appconf==1.0.5
django-cas-ng==4.3.0
django-image-cropping==1.7
django-paypal==2.0
easy-thumbnails==2.8.3
idna==3.3
josepy==1.13.0
lxml==4.9.1
mozilla-django-oidc==2.0.0
paypal==1.2.5
Pillow==9.2.0
psycopg==3.0.15
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.1.0
python-cas==1.6.0
python-decouple==3.6
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
twilio==7.12.0
tzdata==2022.1
urllib3==1.26.11
=======
asgiref==3.5.2
Authlib==1.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.1
Django==4.0.5
django-appconf==1.0.5
django-cas-ng==4.3.0
django-image-cropping==1.7
django-paypal==2.0
easy-thumbnails==2.8.3
idna==3.3
josepy==1.13.0
lxml==4.9.1
mozilla-django-oidc==2.0.0
paypal==1.2.5
Pillow==9.2.0
psycopg==3.0.15
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.1.0
python-cas==1.6.0
python-decouple==3.6
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
twilio==7.12.0
tzdata==2022.1
urllib3==1.26.11
>>>>>>> 0503b05fa942279373ff5a0f0f2025b27e9e34bf
